Fix outline scroll behaviour
scrollIntoView() carries the potential to scroll the entire viewport. With block:nearest this doesn't happen but then the scroll behaviour is worse. Setting position:fixed on body breaks other parts of the layout. Manually scrolling to the centre like this is the best option.

Also fixes horizontal scroll, only do so if needed.